Refactoring and adding functionality
Refactored much of the file which was adding stuff to hooks manually
before to make use of use-package functionality for that which seems
to make it cleaner. There was also some redundant stuff, as well as
typescript-ts-mode and eglot which were configured in a weird way, so
I changed that.

The added functionality is simple, when in org files I often wanted to
be able to count words I write so I added a quick command for
that. The command takes the selected region, splits it in words and
prints the length of the list. I don't know if this scales to longer
texts, but that's something I'll think about later.