Disable an extension and remove extraneous parens
The biggest change is removing the exec-path-from-shell package, which
I seem to have added back in 2023 and which might not serve much
purpose anymore. I marked it as disabled in case I need it after all,
but I honestly doubt it. It was the package which Emacs spent the most
time on while starting up, so this significantly improves the startup
time.